refact(VocabularyCell): 1. View - ViewModel 분리 2. 함수 네이밍 수정
close : #9
